+++ tests/testthat/test-expectation.R
@@ -12,3 +12,17 @@ test_that("expect warns if no `failure_message`", {
 test_that("info only evaluated on failure", {
   expect_error(expect(TRUE, "fail", info = stop("!")), NA)
 })
+
+test_that("can subclass expectation", {
+  exp <- new_expectation("failure", "didn't work", .subclass = "foo", bar = "baz")
+  expect_true(inherits_all(exp, c("foo", "expectation_failure", "expectation", "error", "condition")))
+  expect_identical(attr(exp, "bar"), "baz")
+})
+
+test_that("`expect()` and `exp_signal()` signal expectations", {
+  expect_error(expect(TRUE, ""), regexp = NA)
+  expect_error(expect(FALSE, ""), class = "expectation_failure")
+
+  expect_error(exp_signal(new_expectation("success", "")), regexp = NA)
+  expect_error(exp_signal(new_expectation("failure", "")), class = "expectation_failure")
+})